python/requirements.txt
python-dotenv
agent-framework==1.0.0b260123
agent-framework-azure-ai==1.0.0b260123
agent-framework-mlx==0.4.1
azure-ai-projects==2.0.0b3
azure-ai-agents==1.2.0b5
opentelemetry-semantic-conventions-ai==0.4.13
pydantic>=2.0.0

agent-framework-mlx==0.4.1       # Apple Silicon (MLX) — default
transformers==4.50.0             # Cross-platform (Transformers)
torch>=2.0.0                     # Required by transformers backend
accelerate>=0.26.0               # Required for device_map support